Refactor storage related issues.
