+++ src/main/java/seedu/address/MainApp.java
@@ -19,13 +19,10 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.legacy.AddressBook;
 import seedu.address.model.legacy.ReadOnlyAddressBook;
 import seedu.address.model.util.SampleDataUtil;
-import seedu.address.storage.AddressBookStorage;
 import seedu.address.storage.CentralManager;
-import seedu.address.storage.JsonAddressBookStorage;
-import seedu.address.storage.JsonManagerStorage;
+import seedu.address.storage.JsonCentralManagerStorage;
 import seedu.address.storage.JsonUserPrefsStorage;
 import seedu.address.storage.Storage;
 import seedu.address.storage.StorageManager;
@@ -59,8 +56,8 @@ public class MainApp extends Application {
         UserPrefsStorage userPrefsStorage = new JsonUserPrefsStorage(config.getUserPrefsFilePath());
         UserPrefs userPrefs = initPrefs(userPrefsStorage);
         //AddressBookStorage addressBookStorage = new JsonAddressBookStorage(userPrefs.getAddressBookFilePath());
-        JsonManagerStorage jsonManagerStorage = new JsonManagerStorage(userPrefs.getAddressBookFilePath());
-        storage = new StorageManager(jsonManagerStorage, userPrefsStorage);
+        JsonCentralManagerStorage jsonCentralManagerStorage = new JsonCentralManagerStorage(userPrefs.getAddressBookFilePath());
+        storage = new StorageManager(jsonCentralManagerStorage, userPrefsStorage);
 
         initLogging(config);
